adding basic tcp_server
